fix : /user 라우터 테스트 수정
