+++ server/__test__/service.test.ts
@@ -26,20 +26,6 @@ describe("서비스 계층 테스트", () => {
     sequelize.close();
   });
 
-  describe("지갑주소를 이용하여 사용자의 이름을 가지고 오는 서비스", () => {
-    test("존재하는 지갑 주소", async () => {
-      const query = { walletAddress: keys.test.USER_WALLET };
-      const user = await findUserNameByWalletAddress(query);
-      expect(user?.name).toBe("김퀵커");
-    });
-
-    test("존재하지 않는 지갑 주소", async () => {
-      const query = { walletAddress: "아무개" };
-      const user = await findUserNameByWalletAddress(query);
-      expect(user).toBe(null);
-    });
-  });
-
   describe("출발지와 도착지의 주소 좌표를 데이터베이스에서 가지고 오는 서비스", () => {
     test("존재하는 오더 넘버 입력", async () => {
       const expectResult =  {
