+++ server/routes/Room.ts
@@ -6,9 +6,6 @@ const router = express.Router();
 
 // GET /room
 router.get("/", OrderController.getRoomInfo);
-/** 
- * @TODO : GET /room/message 테스트 코드 작성 필요
- */
 // GET /room/message
 router.get("/message", ChatController.getRecentMessageInfo);
